refactor buf modules
